+++ pytest_picked/plugin.py
@@ -1,3 +1,4 @@
+from fnmatch import fnmatch
 import _pytest
 
 from .modes import Branch, Unstaged
@@ -7,9 +8,15 @@ def pytest_addoption(parser):
     group = parser.getgroup("picked")
     group.addoption(
         "--picked",
-        action="store_true",
+        action="store",
         dest="picked",
-        help="Run the tests related to the changed files",
+        choices=("only", "first"),
+        nargs="?",
+        const="only",
+        help=(
+            "Run the tests related to the changed files either on their own, "
+            "or first"
+        ),
     )
     group.addoption(
         "--mode",
@@ -21,11 +28,7 @@ def pytest_addoption(parser):
     )
 
 
-def pytest_configure(config):
-    picked_plugin = config.getoption("picked")
-    if not picked_plugin:
-        return
-
+def _get_affected_paths(config):
     picked_mode = config.getoption("picked_mode")
     test_file_convention = config._getini(  # pylint: disable=W0212
         "python_files"
@@ -41,12 +44,39 @@ def pytest_configure(config):
         error = "Invalid mode. Options: `{}`.".format(", ".join(modes.keys()))
         _write(config, [error])
         config.args = []
+        raise ValueError(error)
     else:
-        picked_files, picked_folders = mode.affected_tests()
+        return mode.affected_tests()
 
-        config.args = picked_files + picked_folders
 
-        _display_affected_tests(config, picked_files, picked_folders)
+def pytest_configure(config):
+    picked_type = config.getoption("picked")
+    if not picked_type or picked_type != "only":
+        return
+
+    picked_files, picked_folders = _get_affected_paths(config)
+    config.args = picked_files + picked_folders
+    _display_affected_tests(config, picked_files, picked_folders)
+
+
+def pytest_collection_modifyitems(session, config, items):
+    picked_type = config.getoption("picked")
+    if not picked_type or picked_type != "first":
+        return
+
+    affected_files, affected_folders = _get_affected_paths(config)
+    match_paths = affected_files + affected_folders
+    # only reorder if there was anything matched
+    if match_paths:
+        run_first = []
+        run_later = []
+        for item in items:
+            item_path = item.location[0]
+            if any(fnmatch(item_path, m) for m in match_paths):
+                run_first.append(item)
+            else:
+                run_later.append(item)
+        items[:] = run_first + run_later
 
 
 def _display_affected_tests(config, files, folders):